pyproject.toml
[project]
name = "keepsultan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.3.0",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.15.0",
]

[project.scripts]
keep-sultan = "KeepSultanGUI:main"
